The new GUI is almost finished
* Deleted old GUI;
* Deleted unuseful "tools.py" file;
* Code readability has been increased.
