+++ GUI_Parts/text_box.py
@@ -1,4 +1,4 @@
-from GUI.label import Label
+from GUI_Parts.label import Label
 
 from pygame import KEYDOWN, K_KP_ENTER, K_RETURN, K_BACKSPACE, MOUSEBUTTONDOWN
 from pygame.time import get_ticks
